create types module
