+++ vercre-holder/src/issuance/offer.rs
@@ -11,7 +11,7 @@ use tracing::instrument;
 use uuid::Uuid;
 use vercre_core::error::Err;
 use vercre_core::provider::CredentialConfiguration;
-use vercre_core::vci::{CredentialOffer, MetadataRequest};
+use vercre_core::types::issuance::{CredentialOffer, MetadataRequest};
 use vercre_core::{err, Result};
 
 use crate::issuance::{Issuance, Status};
