+++ vercre-holder/tests/issuance.rs
@@ -1,7 +1,7 @@
 use assert_let_bind::assert_let;
 use insta::assert_yaml_snapshot as assert_snapshot;
 use providers::issuance::{Provider as IssuanceProvider, Provider, CREDENTIAL_ISSUER, NORMAL_USER};
-use vercre_core::vci::CreateOfferRequest;
+use vercre_core::types::issuance::CreateOfferRequest;
 
 // This tests the entire issuance flow end-to-end. See the issuance module for stepwise tests.
 #[tokio::test]
