+++ vercre-issuer/src/lib.rs
@@ -103,8 +103,8 @@ use vercre_core::provider::{
     Callback, ClientMetadata, IssuerMetadata, ServerMetadata, StateManager, Subject,
 };
 // TODO: move Claims into jwt module
-pub use vercre_core::vci::{GrantType, ProofClaims};
-pub use vercre_core::{ provider, Result};
+pub use vercre_core::types::issuance::{GrantType, ProofClaims};
+pub use vercre_core::{provider, Result};
 use vercre_vc::proof::Signer;
 
 /// Endpoint is used to surface the public Verifiable Presentation endpoints to
