+++ vercre-verifier/src/create_request.rs
@@ -49,7 +49,9 @@ use uuid::Uuid;
 use vercre_core::error::Err;
 use vercre_core::provider::{Callback, ClientMetadata, StateManager};
 #[allow(clippy::module_name_repetitions)]
-pub use vercre_core::vp::{CreateRequestRequest, CreateRequestResponse, DeviceFlow, RequestObject};
+pub use vercre_core::types::presentation::{
+    CreateRequestRequest, CreateRequestResponse, DeviceFlow, RequestObject,
+};
 use vercre_core::{err, gen, Result};
 use vercre_exch::{Format, PresentationDefinition};
 use vercre_vc::proof::{Algorithm, Signer};
