+++ vercre-verifier/src/request.rs
@@ -20,7 +20,7 @@ use tracing::instrument;
 use vercre_core::error::Err;
 use vercre_core::provider::{Callback, ClientMetadata, Signer, StateManager};
 #[allow(clippy::module_name_repetitions)]
-pub use vercre_core::vp::{RequestObjectRequest, RequestObjectResponse};
+pub use vercre_core::types::presentation::{RequestObjectRequest, RequestObjectResponse};
 use vercre_core::{err, jws, Result};
 
 use super::Endpoint;
@@ -99,7 +99,7 @@ where
 mod tests {
     use insta::assert_yaml_snapshot as assert_snapshot;
     use providers::presentation::{Provider, VERIFIER};
-    use vercre_core::vp::RequestObject;
+    use vercre_core::types::presentation::RequestObject;
 
     use super::*;
