+++ vercre-verifier/src/state.rs
@@ -4,7 +4,7 @@ use chrono::{DateTime, TimeDelta, Utc};
 use derive_builder::Builder;
 use serde::{Deserialize, Serialize};
 use vercre_core::error::Err;
-use vercre_core::vp::RequestObject;
+use vercre_core::types::presentation::RequestObject;
 use vercre_core::{err, Result};
 
 pub enum Expire {
